+++ manager/docs/CHANGELOG.md
@@ -8,5 +8,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- Initial release of the ESUP Runner Manager.
 
+- Initial release of the ESUP Runner Manager.